requirements.txt
# Core Django Dependencies
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Environment & Configuration
django-environ==0.11.2
python-decouple==3.8

# Authentication & Security
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.0

# Task Queue & Background Jobs
celery==5.3.4
redis==5.0.1
django-celery-beat==2.6.0encies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Environment & Configuration
django-environ==0.11.2
python-decouple==3.8

# Authentication & Security
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.0

# Task Queue & Background Jobs
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.0

# Notifications
twilio==8.10.0

# Payments
stripe==7.8.0

# Image & File Storage
cloudinary==1.36.0
django-cloudinary-storage==0.3.0

# UI & Frontend
django-bootstrap5==23.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3
